Modify build script

+++ specials/make_special_x86_64.sh
@@ -4,6 +4,17 @@ ARCH=x86_64
 
 rm -r ${ARCH}
 ../exec/${ARCH}/bin/nrnivmodl ../mod
-rm ./${ARCH}/hh_k.c
-cp ./hh_k7_3.c ${ARCH}/hh_k.c
+if [ $# -eq 0 ]
+then
+    echo "bench"
+else
+    if [ $1 == 'True' ]
+    then
+        rm ./${ARCH}/hh_k.c
+        cp ~/genie/genie/transpiler/tmp/hh_k.c ${ARCH}/hh_k.c
+    else
+        rm ./${ARCH}/hh_k.c
+        cp ./hh_k7_3.c ${ARCH}/hh_k.c
+    fi
+fi
 ../exec/${ARCH}/bin/nrnivmodl ../mod
